Code/requirements.txt
aniso8601==8.0.0
azure-cognitiveservices-language-textanalytics==0.2.0
azure-common==1.1.24
azure-storage-blob==2.1.0
azure-storage-common==2.1.0
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
cryptography==2.8
dnspython==1.16.0
Flask==1.1.1
Flask-HTTPAuth==3.3.0
Flask-RESTful==0.3.8
idna==2.8
isodate==0.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
msrest==0.6.11
mysql-connector-python==8.0.19
numpy==1.18.1
oauthlib==3.1.0
pandas==1.0.1
protobuf==3.6.1
pycparser==2.19
python-dateutil==2.8.1
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
six==1.14.0
urllib3==1.25.8
Werkzeug==1.0.0
